accelerator/imagenet_demo/requirements.txt
PyYAML >= 5.4.1
scipy